DCP support for Sync Gateway
Support for using the Couchbase Server DCP feed instead of the TAP feed when available.  Uses the cbdatasource library from go-couchbase for interaction with the DCP feed, which handles cluster rebalancing and bucket failover.  Recommendation from Steve Yen is that cbdatasource needs to manage it's own connection, so we're not reusing SG's existing bucket connection for the DCP feed.  This requires that bucket auth information is passed to cbdatasource when initializing the DCP feed, so we're now saving the bucket spec along with the bucket to have that available on feed startup.

For standard feed usage, SG will make a stats vbseqno call to the server to identify the high sequence number in each bucket, and use that to start a DCP feed listening for new revisions.

For shadowing usage, SG starts from seq=0 and processes all data on startup. An enhancement to support resuming from the latest is possible by persisting DCP sequence info on snapshot end/start, but I've got concerns about whether the performance impact to the shadower feed of persisting sequence information on every vbucket snapshot.  A similar problem is being discussed in the context of #525 (how to efficiently persist the latest sequence information for all vbuckets), so I plan to revisit once that's completed.

By default, SG will attempt to use a DCP feed when working with a Couchbase bucket.  If DCP isn't supported for the bucket, it reverts to the TAP feed.  To force usage of TAP, users can set the property 'feed_type':'tap' in the database definition in the sync gateway config file.